requirements.txt
Django==1.6
Pillow==2.2.1
cffi==0.8
django-appconf==0.6
django-imagekit==3.0.4
greenlet==0.4.0
pilkit==1.1.5
readline==6.2.4.1
six==1.4.1
wsgiref==0.1.2
